refactor(demo): move add-to-cart-notification to cart directory
