+++ apps/demo/src/app/cart/components/add-to-cart-notification/components/product-added/product-added.component.scss
