+++ apps/demo/src/app/product/pages/product-view/product-view.module.ts
@@ -7,8 +7,8 @@ import { DaffLoadingIconModule } from '@daffodil/design';
 
 import { ProductComponentModule } from '../../components/product/product.module';
 import { AddToCartModule } from '../../components/add-to-cart/add-to-cart.module';
-import { AddToCartNotificationModule } from '../../../modals/add-to-cart-notification/add-to-cart-notification.module';
 import { ProductViewComponent } from './product-view.component';
+import { AddToCartNotificationModule } from '../../../cart/components/add-to-cart-notification/add-to-cart-notification.module';
 
 @NgModule({
   imports: [
